Booster elevator test
This is the snapshot of the booster code as used for the elevator test.
We have confirmed the following:

1. System arms when crossing the `ARM_ALT` altitude.
2. System enters ascent mode when crossing `LAUNCH_FLOOR` altitude.
3. Separation timer only starts when #2 above happens
4. Separation timer causes relay-fire on the relay bank labeled "S".
5. Apogee detection function, descent state is only entered when
   falling from apogee by `DESC_DROP_THRESH` ft.
6. When in descent state, when altimeter detects current_alt <=
   `MAIN_DEPLOYMENT_ALTITUDE`, the relay bank for the main parachute
   fires (marked `R`).

Note for data analysis: the time of parachute and separation relay
actuation is logged to the "flags" field of the data log. This is a bit
field with the following structure:

MSB 0 0 0 0 0 0 a b LSB
                | |
                | Staging charge fired
                Main parachute separation charge fired

All other flight log fields are individual columns in CSV format, and
the CSV file will always be written with the data description header.
Please note that each boot seems to append to the existing Flight.csv
file, so the correct log will have to be extracted from the file. Each
flight will have a _new CSV header_.

The bay implements the following beep codes:

1. One LONG beep right at powerup.
2. Three SHORT beeps once calibration completes.

Long beeps after the power up beep indicate some sort of critical system
failure:

- 3 long beeps indicates that the SD card was not installed.
  More rarely this can also indicate that the SD card is failing to
  initialize, at which point the card should be checked to ensure that
  there is actually a filesystem on it.
- 4 long beeps indicates that the flight log failed to open.

In either of these cases the flight controller will shutdown and will
NEED TO BE RESTARTED after fixing the problem. Please note that the
booster buzzer is a single-pitch buzzer, and only indicates status
through the number and duration of beeps.

Arm-hysteresis is achieved by having two low thresholds: the `ARM_ALT`
and the `LAUNCH_FLOOR`. When the controller detects that the `ARM_ALT`
has been crossed it enters a "detect launch" state, where it will
transition into the ascent phase if the `LAUNCH_FLOOR` is crossed. If
consecutive altitude readings are below the `ARM_ALT` while the state
machine is in the detect launch state, the system will assume that the
rocket was falsely armed, and return to the disarmed state. Subsequent
arm conditions can return the controller to the arm state, i.e. this is
not a locket condition.

The `AVG_CNT` parameter is used to determine the number of samples used
to smooth altimeter data. This has been set to 4 samples, which comes
out to about one to one-and-a-half seconds of averaging, which strikes a
good balance between deployment latency and noise suppression. Please
note that parachute deployment will typically by delayed by the sum of
this latency and the ematch ignition latency, so plan for around one
second of latency (i.e. extra time spent free-falling) while dailing in
the parachute heights.

The staging charge and separation charge are both on a paired
ACTIVE-LOW relay bank. There is a very specific sequence to enable the
digital pins that output to such safety-critical active low devices
since Arduino pins start out in LOW output mode when the pinMode is
first set to output by default (due to the pin bank register starting at
zero during cold boot):

1. `digitalWrite()` to BOTH active low relays pins BEFORE setting the
   pinmode to output. Doing this after means that there will be a short,
   undetermined, period of time between the toggle of the pin to output
   and the digitalWrite that sets it low while the relay is ON. This is
   very bad. Note writing pin state before enabling output still works,
   since digitalWrite just sets the corresponding pin bit in the pin
   bank control register, so the pin will then startup in the correct
   state.
2. Set `pinMode` to output.

WARNING: THIS IS ELEVATOR TEST FIRMWARE. EDIT THE CONSTANT BLOCK AT THE
TOP OF THE `boosterFSM` FILE TO REFLECT FLIGHT PARAMS BEFORE LAUNCH AND
COMMIT TO GITHUB FOR RECORDS.