Sustainer elevator test
Verified the following requirements for the sustainer elevator test:

1. System arms at `ARM_ALT` and enters ready-to-detect launch mode with
   hysterysis
2. Igniter relay only fires when the rocker is within 20 degrees of
   vertical
3. Igniter relay fires on a timer triggered when the system arms at the
   arm altitude
4. Ascent and descent phases are squenced correctly
5. During descent phase, the drogue fires first, followed by the main
   parachute relay.
6. Parachute relays are always fired, even if the igniter never fired.

The flags field in the flight log is a bit field with the following
format:

MSB 0 0 0 0 a b c d LSB
            | | | |
            | | | Streaming camera subsystem power on
            | | Sustainer ignition charge fired
            | Drogue parachute sep charge fired
            Main parachute sep charge fired

Flags directly correspond to the corresponding digital pin on the
controller (i.e. they directly reflect the state of a relay).
All other flight log fields are individual columns in CSV format, and
the CSV file will always be written with the data description header.
Please note that each boot seems to append to the existing Flight.csv
file, so the correct log will have to be extracted from the file. Each
flight will have a _new CSV header_.

The bay implements the following beep codes:

1. One long beep at between B5/C6 right after powerup.
2. An ASCENDING SCALE starting from A4 once calibration completes.

Long beeps after the power up beep indicate some sort of critical system
failure:

- 3 long beeps at 300Hz indicates that the SD card was not installed.
  More rarely this can also indicate that the SD card is failing to
  initialize, at which point the card should be checked to ensure that
  there is actually a filesystem on it.
- 3 long beeps at 500Hz indicates that the flight log failed to open.

In either of these cases the flight controller will shutdown and will
NEED TO BE RESTARTED after fixing the problem.

Arm-hysteresis is achieved by having two low thresholds: the `ARM_ALT`
and the `LAUNCH_FLOOR`. When the controller detects that the `ARM_ALT`
has been crossed it enters a "detect launch" state, where it will
transition into the ascent phase if the `LAUNCH_FLOOR` is crossed. If
consecutive altitude readings are below the `ARM_ALT` while the state
machine is in the detect launch state, the system will assume that the
rocket was falsely armed, and return to the disarmed state. Subsequent
arm conditions can return the controller to the arm state, i.e. this is
not a lockout condition.

The `AVG_CNT` parameter is used to determine the number of samples used
to smooth altimeter and linear accelerometer data. This has been set to
4 samples, which comes out to about one to one-and-a-half seconds of
averaging, which strikes a good balance between deployment latency and
noise suppression. Please note that parachute deployment will typically
by delayed by the sum of this latency and the ematch ignition latency,
so plan for around one second of latency (i.e. extra time spent
free-falling) while dialing in the parachute heights.

Both separation charges on a paired ACTIVE-LOW relay bank. There is a
very specific sequence to enable the digital pins that output to such
safety-critical active low devices since Arduino pins start out in LOW
output mode when the pinMode is first set to output by default (due to
the pin bank register starting at zero during cold boot):

1. `digitalWrite()` to BOTH active low relays pins BEFORE setting the
   pinmode to output. Doing this after means that there will be a short,
   undetermined, period of time between the toggle of the pin to output
   and the digitalWrite that sets it low while the relay is ON. This is
   very bad. Note writing before enablign output still works, since
   digitalWrite just sets the corresponding pin bit in the pin bank
   control register, so the pin will then startup in the correct state.
2. Set `pinMode` to output.

WARNING: THIS IS ELEVATOR TEST FIRMWARE. THE FLIGHT PARAMETERS AT THE
TOP OF THE FIRMWARE FILE WILL NEED TO BE EDITED AND COMMITTED BACK TO
GITHUB, AND THE PROPER FLIGHT FIRMWARE FLASHED TO THE CONTROLLER BEFORE
LAUNCH.